Move shared web/desktop UI components to core
