+++ packages/core/package.json
@@ -46,6 +46,7 @@
     "json-stable-stringify": "^1.0.1",
     "object.fromentries": "^2.0.0",
     "pako": "^1.0.10",
+    "react-measure": "^2.3.0",
     "react-sanitized-html": "^2.0.0",
     "sanitize-html": "^1.20.1",
     "shortid": "^2.2.13",
