+++ packages/core/ui/App.js
@@ -6,11 +6,11 @@ import Slide from '@material-ui/core/Slide'
 import { makeStyles } from '@material-ui/core/styles'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
-import { inDevelopment } from '@gmod/jbrowse-core/util'
 import { observer, PropTypes } from 'mobx-react'
 import ReactPropTypes from 'prop-types'
 import React, { useEffect } from 'react'
 import { withContentRect } from 'react-measure'
+import { inDevelopment } from '../util'
 import DevTools from './DevTools'
 import Drawer from './Drawer'
