+++ packages/core/ui/NewSessionCards.js
@@ -1,7 +1,6 @@
-import { readConfObject } from '@gmod/jbrowse-core/configuration'
-import CardContent from '@material-ui/core/CardContent'
 import Button from '@material-ui/core/Button'
 import Card from '@material-ui/core/Card'
+import CardContent from '@material-ui/core/CardContent'
 import Container from '@material-ui/core/Container'
 import FormControl from '@material-ui/core/FormControl'
 import Icon from '@material-ui/core/Icon'
@@ -13,6 +12,7 @@ import Typography from '@material-ui/core/Typography'
 import { PropTypes as MobxPropTypes } from 'mobx-react'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
+import { readConfObject } from '../configuration'
 import svInspectorIcon from './sv inspector icon.png'
 
 const useStyles = makeStyles(theme => ({
