+++ packages/core/ui/ResizeHandle.js
