+++ packages/core/ui/index.ts
@@ -1,1 +1,7 @@
+export { default as theme } from './theme'
+export { default as App } from './App'
 export { default as FileSelector } from './FileSelector'
+export { default as PrerenderedCanvas } from './PrerenderedCanvas'
+export { default as ResizeHandle } from './ResizeHandle'
+export { default as StartScreen } from './StartScreen'
+export { default as Tooltip } from './Tooltip'
