+++ packages/core/ui/sv inspector icon.png
