+++ packages/jbrowse-desktop/src/JBrowse.js
@@ -1,4 +1,5 @@
 import '@gmod/jbrowse-core/fonts/material-icons.css'
+import { App, StartScreen, theme } from '@gmod/jbrowse-core/ui'
 import CircularProgress from '@material-ui/core/CircularProgress'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import { ThemeProvider } from '@material-ui/styles'
@@ -8,9 +9,6 @@ import React, { useEffect, useState } from 'react'
 import 'typeface-roboto'
 import { useDebounce } from '@gmod/jbrowse-core/util'
 import rootModel from './rootModel'
-import App from './ui/App'
-import StartScreen from './ui/StartScreen'
-import Theme from './ui/theme'
 
 const { electron = {} } = window
 const { desktopCapturer, ipcRenderer } = electron
@@ -112,7 +110,7 @@ export default observer(() => {
   }
 
   return (
-    <ThemeProvider theme={Theme}>
+    <ThemeProvider theme={theme}>
       <CssBaseline />
       {DisplayComponent}
     </ThemeProvider>
