+++ packages/jbrowse-web/package.json
@@ -47,7 +47,6 @@
     "raf": "^3.4.1",
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
-    "react-measure": "^2.3.0",
     "react-sizeme": "^2.6.7",
     "rxjs": "6.5.2",
     "typeface-roboto": "^0.0.54",
