+++ packages/jbrowse-web/src/JBrowse.js
@@ -1,5 +1,6 @@
 import { readConfObject } from '@gmod/jbrowse-core/configuration'
 import '@gmod/jbrowse-core/fonts/material-icons.css'
+import { App, theme } from '@gmod/jbrowse-core/ui'
 import {
   toUrlSafeB64,
   fromUrlSafeB64,
@@ -15,8 +16,6 @@ import { UndoManager } from 'mst-middlewares'
 import React, { useEffect, useState } from 'react'
 import 'typeface-roboto'
 import rootModel from './rootModel'
-import App from './ui/App'
-import Theme from './ui/theme'
 
 export default observer(({ config, initialState }) => {
   const [status, setStatus] = useState('loading')
@@ -178,7 +177,7 @@ export default observer(({ config, initialState }) => {
     DisplayComponent = <App session={root.session} />
 
   return (
-    <ThemeProvider theme={Theme}>
+    <ThemeProvider theme={theme}>
       <CssBaseline />
       {DisplayComponent}
     </ThemeProvider>
