--- packages/jbrowse-web/src/ui/theme.js
@@ -1,9 +1,0 @@
-import { createMuiTheme } from '@material-ui/core/styles'
-
-export default createMuiTheme({
-  palette: {
-    // type: 'dark',
-    primary: { main: '#396494' },
-    secondary: { main: '#399469' },
-  },
-})
